+++ src/components/auth/common/AuthForm/index.tsx
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+const AuthForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  margin-top: 20px;
+
+  & > label {
+    margin-bottom: 0.8rem;
+  }
+`;
+
+export default AuthForm;
