+++ src/components/auth/register/RegisterForm/style.ts
@@ -1,16 +1,5 @@
 import styled from '@emotion/styled';
 
-export const RegisterFormRoot = styled.form`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin-top: 20px;
-
-  & > label {
-    margin-bottom: 0.8rem;
-  }
-`;
-
 export const RegisterButtonContainer = styled.div`
   display: flex;
   gap: 0.5rem;
