+++ src/components/auth/verify/CodeCheck/style.ts
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion';
 import styled from '@emotion/styled';
 
-export const CodeDiv = styled(motion.form)`
+export const CodeForm = styled(motion.form)`
   display: flex;
   flex-direction: column;
   gap: 1rem;
